+++ test/tests/TestInputConfig.java
@@ -1,23 +1,24 @@
 package tests;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Test-only in-memory input configuration Used to load/read data from
+ * Test-only in-memory input configuration used to load/read data.
  */
 public class TestInputConfig {
-  List<Integer> inputData;
+  List<BigInteger> inputData;
 
   public TestInputConfig() {
     this.inputData = new ArrayList<>();
   }
 
-  public TestInputConfig(List<Integer> inputData) {
+  public TestInputConfig(List<BigInteger> inputData) {
     this.inputData = inputData;
   }
 
-  public void addInputData(Integer value) {
+  public void addInputData(BigInteger value) {
     this.inputData.add(value);
   }
 }
